+++ src/main/java/ru/startup/controllers/MuseumController.java
@@ -1,6 +1,8 @@
 package ru.startup.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import ru.startup.dto.MuseumDTO;
 import ru.startup.model.entertainment.EntertainmentType;
@@ -17,13 +19,20 @@ public class MuseumController {
     }
 
     @GetMapping("/api/museum/{id}")
-    public MuseumDTO getMuseumById(@PathVariable Long id){
-        return museumService.getMuseumById(id);
+    public ResponseEntity<MuseumDTO> getMuseumById(@PathVariable Long id){
+        if (!museumService.existsById(id)){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        return new ResponseEntity<>(museumService.getMuseumById(id),HttpStatus.OK);
     }
 
     @DeleteMapping("/api/museum/{id}")
-    public void deleteMuseumById(@PathVariable Long id){
+    public ResponseEntity<Void> deleteMuseumById(@PathVariable Long id){
+        if (!museumService.existsById(id)){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         museumService.deleteMuseumById(id);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @PostMapping("/api/museum")
