+++ src/main/java/ru/startup/controllers/PotteryController.java
@@ -1,6 +1,8 @@
 package ru.startup.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import ru.startup.dto.PotteryDTO;
 import ru.startup.model.entertainment.EntertainmentType;
@@ -17,13 +19,20 @@ public class PotteryController {
     }
 
     @GetMapping("/api/pottery/{id}")
-    public PotteryDTO getPotteryById(@PathVariable Long id){
-        return potteryService.getPotteryById(id);
+    public ResponseEntity<PotteryDTO> getPotteryById(@PathVariable Long id){
+        if (!potteryService.existsById(id)){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        return new ResponseEntity<>(potteryService.getPotteryById(id), HttpStatus.OK);
     }
 
     @DeleteMapping("/api/pottery/{id}")
-    public void deletePotteryById(@PathVariable Long id){
+    public ResponseEntity<Void> deletePotteryById(@PathVariable Long id){
+        if (!potteryService.existsById(id)){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         potteryService.deletePotteryById(id);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @PostMapping("/api/pottery")
