+++ src/main/java/ru/startup/controllers/ShootingClubController.java
@@ -1,6 +1,8 @@
 package ru.startup.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import ru.startup.dto.ShootingClubDTO;
 import ru.startup.model.entertainment.EntertainmentType;
@@ -17,13 +19,20 @@ public class ShootingClubController {
     }
 
     @GetMapping("/api/shooting-club/{id}")
-    public ShootingClubDTO getShootingClubById(@PathVariable Long id){
-        return shootingClubService.getShootingClubById(id);
+    public ResponseEntity<ShootingClubDTO> getShootingClubById(@PathVariable Long id){
+        if (!shootingClubService.existsById(id)){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        return new ResponseEntity<>(shootingClubService.getShootingClubById(id), HttpStatus.OK);
     }
 
     @DeleteMapping("/api/shooting-club/{id}")
-    public void deleteShootingClubById(@PathVariable Long id){
+    public ResponseEntity<Void> deleteShootingClubById(@PathVariable Long id){
+        if (!shootingClubService.existsById(id)){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         shootingClubService.deleteShootingClubById(id);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @PostMapping("/api/shooting-club")
