+++ src/main/java/ru/startup/controllers/VRClub.java
@@ -1,6 +1,8 @@
 package ru.startup.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import ru.startup.dto.VRClubDTO;
 import ru.startup.model.entertainment.EntertainmentType;
@@ -17,13 +19,20 @@ public class VRClub {
     }
 
     @GetMapping("/api/vr-club/{id}")
-    public VRClubDTO getVRClubById(@PathVariable Long id){
-        return vrClubService.getVRClubById(id);
+    public ResponseEntity<VRClubDTO> getVRClubById(@PathVariable Long id){
+        if (!vrClubService.existsById(id)){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        return new ResponseEntity<>(vrClubService.getVRClubById(id), HttpStatus.OK);
     }
 
     @DeleteMapping("/api/vr-club/{id}")
-    public void deleteVRClubById(@PathVariable Long id){
+    public ResponseEntity<Void> deleteVRClubById(@PathVariable Long id){
+        if (!vrClubService.existsById(id)){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         vrClubService.deleteVRClubById(id);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @PostMapping("/api/vr-club")
